fileserve: wait arbitrary time when it says so, only return when explicitely asked
git-svn-id: https://plowshare.googlecode.com/svn/trunk@1383 51fab9ec-ea3e-11dd-80ed-bda8ba2650cd

+++ src/modules/fileserve.sh
@@ -76,8 +76,13 @@ fileserve_download() {
 
         elif match 'timeLimit' "$JSON1"; then
             log_error "time limit, you must wait"
-            rm -f $COOKIES
-            return 1
+            if test "$NOARBITRARYWAIT"; then
+                rm -f $COOKIES
+                return 1
+            fi
+
+            wait $STOP_FLOODING seconds || return 2
+            continue
 
         elif ! match 'success' "$JSON1"; then
             log_error "unexpected error, site update?"
